move additional files into subfolder layouts
